feat: reusable capsuleview
